Extract history-entry. Add html and unknown repl entry types
Preparing to load html and other types of commands from a file for the lessons.

Try in the repl:
    (with-meta {:value "<h1>oh HI</h1><h3>How are you doing!</h3>"} {:type :html})

If a response has meta with a type, it'll be rendered with special renderer if
it is a known type.

If the value of the entry is a map with the `:value` key, that'll be the
contents that get rendered (may bite us in the future but it's only for values
with meta so shouldn't be a huge problem).